test: modified pr_test and moved fscloud to other test
